requirements.txt
# eJARN 수집·정규화 프로젝트
requests>=2.28.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
readability-lxml>=0.8.1
playwright>=1.40.0

# LLM 요약/분류 및 에이전트
openai>=1.0.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
streamlit>=1.38.0
